server/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
Pillow==10.0.0
pytesseract==0.3.10
opencv-python==4.8.0.74
numpy==1.24.3
python-dotenv==1.0.0